[stub] fix(mount): remove problematic reduction of in-word prefix
With the current implementation, the originally specified options
vanish on the tab completion when they are quoted.  For example, with
the following example, 'abort' will be removed.

$ mount -o 'abort',a[TAB]

The variable `prev` in this context is originally extracted from
`cur`, which is the word after splitting by COMP_WORDBREAKS.  This
means that an additional splitting should not be needed.  Even if
special characters are included in `prev`, they must be quoted in
order to appear there so should not be considered wordbreaks.

The code was originally introduced in commit db8d3eeee (2011-01-19),
where no context is given.