fix(cd): fix usage of glob matching
This comparison with the quoted pattern "*/" seems to have existed
since the code is first introduce in commit a1969bdd.  The reasoning
is not explained in the original commit message.  The pattern matching
*/ is likely to be intended, but it does not work as expected because
it is quoted.

However, this pattern matching is actually not needed because when the
current word "$cur" matches the pattern */, the current implementation
will never generate itself as a completion but instead generates the
subdirectories under that directory.

In this patch, we just remove the pattern matching.

References:
https://github.com/scop/bash-completion/commit/a1969bdd571590370a0350b3e6ff53be54948344