fix(cd): filter cdable_vars with their values
This commit fixes two problems.

* One problem is that "shopt -q cdable_vars" is checked when the "cd"
  completion is loaded.  When the user changes the setting
  "cdable_vars" after the "cd" completion is loaded, the change is not
  reflected.

* Another problem is that all the variable names are generated when
  "cdable_vars" is turned on.  This becomes a problem when shell
  frameworks definiing many variables are loaded.  The "cd" completion
  would generate many shell variables that are not able/intended to be
  used as "cdable_vars".  This always causes the message "Display all
  *** possibilities (y/n)" in an attempt of the cd command, which is
  not useful.

In this patch, we try to generate the variable names inside the
completion function so that one can switch the generation of
"cdable_vars" by "shopt -q cdable_vars" at the completion time.  We
also filter the variable names to generate only the variable names
containing a valid directory path.