fix(_comp_load): properly handle backslashed cmd
In the current "main" branch, when we a completion setting is found,
if the originally requested command is a backslashed one, we try to
copy the completion setting to the backslashed one using
"$backslash$cmd".

However, this does not work.  "$backslash$cmd" is not the command name
originally specified by the user because $cmd contains the absolute
path to the command after PATH search.  This comes from the confusion
on the variable $cmd, as implied by the wrong description of $cmd in
the code comment on the previous line.  In the current implementation,
"cmd" is 1) first initialized to be the original command "$1", 2) then
updated to be the one without backslash, and 3) finally overwritten by
the absolute path while saving its old value to "origcmd".
"$backslash$cmd" receives the result of 3 for $cmd while it expects
the result of 2, which is the cause of the present problem.

In this patch, we introduce separate variables for 1-3 and avoid
changing the content of the variables dynamically.  Finally, we
specify the saved value of 1 instead of "$backslash$cmd".