fix(_comp_load): fix a regression for backslashed command names
Currently, completions are not loaded for the command names starting
with a backslash.

In commit f6a221148, processing of the beginning backslash in the
command name has been moved to an earlier position in
__load_completion (corresponding to _comp_load currently).  However,
at the same time, the target of the backslash removal was changed from
the variable "cmdname" to the variable "cmd".  This resulted in a
backslash remaining in "cmdname" and searching for completion
filenames including a backslash.

When we update the variable "cmd", the variable "cmdname" should also
be updated. Or, actually, one can move the position of the backslash
processing even earlier.  In this patch, we move the position of the
backslash even earlier, and determine the value of "cmdname" using
"cmd" after the backslash removal.